Move API related classes to the correct packages
