+++ app/src/test/java/com/lima2017/neuralguide/api/web/NeuralGuideDataTest.java
@@ -1,4 +1,4 @@
-package com.lima2017.neuralguide.api;
+package com.lima2017.neuralguide.api.web;
 
 import org.junit.Test;
