+++ app/src/test/java/com/lima2017/neuralguide/api/web/NeuralGuideResultDataTest.java
@@ -1,4 +1,6 @@
-package com.lima2017.neuralguide.api;
+package com.lima2017.neuralguide.api.web;
+
+import com.lima2017.neuralguide.api.web.NeuralGuideResultData;
 
 import org.junit.Test;
